+++ src/main/java/ladder/domain/Line.java
@@ -1,8 +1,6 @@
 package ladder.domain;
 
 import java.util.Objects;
-import ladder.domain.coordinate.Column;
-import ladder.domain.coordinate.Row;
 import ladder.utils.LineStrategy;
 
 public class Line {
