+++ src/test/java/ladder/domain/LadderTest.java
@@ -4,8 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Set;
-import ladder.domain.coordinate.Column;
-import ladder.domain.coordinate.Row;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
